es/utils/requirements.txt
pyyaml
nose